magic

+++ 0x06-python-classes/103-magic_class.py
@@ -9,7 +9,7 @@ class MagicClass:
 
     def __init__(self, radius):
         """ initialize attributes """
-        if type(radius) is not int or type(radius) is not int:
+        if type(radius) is not int or type(radius) is not float:
             raise TypeError("radius must be a number")
         self.__radius = radius
